requirements.txt
streamlit
langchain-google-genai
langchain-chroma
python-dotenv
requests
colorama